Bug #18047796 MTR TEST MAIN.OPENSSL_1 FAILS ON FEDORA 19 WITH OPENSSL 1.0.1E
Description: mtr test main.openssl_1 fails on Fedora 19 x86_64, server is built against
system openssl 1.0.1e-35.

Analysis: The cipher "EDH-RSA-DES-CBC-SHA" is removed from the open_ssl since "openssl-1.0.1e-35" source. Mentioned that its a week cipher. Because of that the test case which is looking for the cipher is not finding in the cipher list.

Fix: We have replaced the cipher check "EDH-RSA-DES-CBC-SHA" with another common "AES256-SHA" cipher in the test case.