Bug#17803141 ASSERT FAILURE WHEN RUNNING OPTIMIZE TABLE              WITH GTID_MODE=ON
Problem:
=======

The rpl_failed_optimize MTR test running with GTIDs enabled is making
an assert to fail at the slave server. This happens when the slave SQL
thread tries to binlog an OPTIMIZE TABLE statement for an inexistent
table that was successfully executed (and binlogged) at the master. 

Background:
==========

It is a general property of all commands based on mysql_admin_table() to
return errors as part of a result set instead of failing the statement
as they rarely end-up in error.

Because of this property it is possible to do something like
"OPTIMIZE TABLE t1, t2, t3;" and get different results/errors for t1,
t2 and t3 and do not fail the whole statement.

Also, from locking perspective, each table in this list is processed
individually: i.e. the client session will acquire locks on a single
table, open it, process it and then release the locks
(and commit/rollback transaction) before processing next table in the
statement.

Analysis:
========

The reported problem happens when the rpl_failed_optimize test is 
running with GITDs enabled and tries to execute an OPTIMIZE TABLE for
an inexistent table in the master server.

This command succeeds at the master, goes to master's binlog, goes to
slave's relay log, is executed by the SQL thread and then fails to be
binlogged to the slave's binlog.

The issue happens inside mysql_admin_table(), that is called by
Sql_cmd_optimize_table::execute().

Sql_cmd_optimize_table::execute() will binlog the current statement
unless the call to mysql_admin_table() returns error.

For each table to be processed, the mysql_admin_table() function
will call open_and_lock_tables() function that might call
trans_rollback_stmt() if an error is found when trying to open an
table (if the table doesn't exists for example).

The trans_rollback_stmt() function will call gtid_rollback() to ensure
that the GTID_NEXT associated with the current session will not be used
in the case of a rolled back statement.

The problem is that mysql_admin_table() has a property of not throwing
errors to the client session as stated earlier in the Background
section. So, because of this property, mysql_admin_table() will return
that it has succeed to Sql_cmd_optimize_table::execute(), even if some
minor errors had happened.

As the mysql_admin_table() call has succeed,
Sql_cmd_optimize_table::execute() will try to binlog the statement
but it will fail to do so because GTIDs are enabled and some minor error
happened (and gtid_rollback() was called), and no GTID_NEXT is available
to binlog the current statement.

The same behavior will happen not only with OPTIMIZE TABLE, but also
with ANALYZE TABLE and REPAIR TABLE table maintenance statements, all
based on mysql_admin_table().

Fix:
===

This fix aims only to ensure that the same behavior that happened at the
master will happen in the slave, even if GTIDs are enabled. It will
assure that if a GTID_NEXT was set to GTID_GROUP for the statement
execution, the same GTID will be available to be binlogged if the
command succeeds.

@ sql/sql_class.h

A new flag was created into THD object named "skip_gtid_rollback".

@ sql/sql_class.cc

This flag is initialized as "false" at the THD object constructor.

@ sql/rpl_gtid_execution.cc

At the gtid_rollback() function, a test in the "skip_gtid_rollback"
flag will prevent the rollback of the GTID if the flag is enabled.

@ sql/sql_admin.cc

In the beginning of the mysql_admin_table() function, we added code
to check if the current statement have to go to binlog, if the
GTID_NEXT type is set to GTID_GROUP and if the THD->skip_gtid_rollback
flag is not enabled, setting the result on a local flag variable named
"gtid_rollback_must_be_skipped".

Then, before entering the main loop for processing the tables, we check 
if the "gtid_rollback_must_be_skipped" flag is set, changing
THD->skip_gtid_rollback to "true" if needed. 

Finally, we check if the "gtid_rollback_must_be_skipped" flag is set,
restoring the gtid_rollback() normal behavior by setting
THD->skip_gtid_rollback to "false" again in three mysql_admin_table()
locations:
- Just before returning that the mysql_admin_table() has succeeded;
- Before the "err:" section calls to trans_rollback_stmt();
- Before mysql_admin_table() exit in the case of a  
  ER_PARTITION_MGMT_ON_NONPARTITIONED error.