Bug#17959689: MAKE GCC AND CLANG GIVE CONSISTENT COMPILATION WARNINGS
Post-push fix: On some OS X configurations we use different compilers for C and
C++ code. The original patch could cause problems for such configurations as
it used the C compiler to check if -Wvla was supported and then applied the flag
to both the C and C++ compiler. So if clang was used as C compiler (supports the flag)
and G++ 4.2 was used as C++ compiler (does not support the flag), the build broke.

This patch fixes the problem by doing the check twice, once for the C compiler
and once for the C++ compiler.