Bug#17661919 ASSERT 0 IN ROW_LOG_TABLE_APPLY_UPDATE(), ALTER TABLE REBUILD WITH BLOBS
There is a problem with the table-rebuilding variant of
online ALTER TABLE when there is concurrent DML of the form:

BEGIN;
INSERT INTO t SET blob_column='old_value';
UPDATE t SET blob_column='new_value';
ROLLBACK;

The ROLLBACK would perform the equivalent of purging the INSERT.
This would generate the following log records:

ROW_T_INSERT
ROW_T_UPDATE
ROW_T_UPDATE -- for the ROLLBACK of the UPDATE
ROW_T_DELETE -- for the ROLLBACK of the INSERT

(The BLOB 'new_value' for the UPDATE would not be purged before the
transaction performing the online ALTER TABLE has been committed or
rolled back, because the purge cannot touch rows that are visible in
the read view of any active transaction.)

Another important consideration is the purge of BLOBs that were made
purgeable by a committing an UPDATE or DELETE operation. While it is
true that online ALTER TABLE will acquire an exclusive table lock
during the preparation phase, it is possible that some transactions
were started before the ALTER TABLE, and these transactions did not
access the table yet. So, we can have concurrent DML like this:

-- BEGIN was before the ALTER TABLE started.
-- The transaction did not access the table until this
-- (otherwise there would have been a MDL conflict with the ALTER):
DELETE FROM t WHERE id=42; -- first access to this table
INSERT INTO t VALUES(101,'old value');
UPDATE t SET blob_column='new value' WHERE id=101;
COMMIT;

This would generate the following log records:

ROW_T_DELETE
ROW_T_INSERT
ROW_T_UPDATE
[purge does not generate log, it only updates log->blobs]

The ROW_T_DELETE is not a problem, because it will include enough BLOB
prefixes to identify records if needed. Only ROW_T_INSERT or
ROW_T_UPDATE may suffer from missing BLOBs.

If we skip the ROW_T_INSERT because of the purged 'old value', we must
interpret the ROW_T_UPDATE as ROW_T_INSERT of (101,'new value'). But,
we must do this in a way that is compatible with the above ROLLBACK
scenario.

When processing ROW_T_INSERT or ROW_T_UPDATE such that the BLOBs are
missing, we can conclude that the BLOBs were removed due to one of:

(1) a later ROLLBACK of an INSERT (there will be ROW_T_DELETE)
(2) running purge for DELETE;COMMIT; (there will be ROW_T_DELETE)
(3) running purge for UPDATE;COMMIT; (there will be ROW_T_UPDATE)

Thus, we can anticipate that there would be a ROW_T_DELETE for
removing the record or a ROW_T_UPDATE for replacing the record. So, we
can safely ignore the ROW_T_INSERT operation, or we can interpret the
ROW_T_UPDATE as ROW_T_DELETE when some BLOBs for the 'after' image are
missing.

When processing ROW_T_UPDATE or ROW_T_DELETE, we must make sure that
the record exists and (in case we are executing ADD PRIMARY KEY) it is
stamped with the same DB_TRX_ID,DB_ROLL_PTR value that was logged in
old_pk.

If there is a DB_TRX_ID,DB_ROLL_PTR mismatch, it means that our record
had already been deleted and we can safely ignore the ROW_T_DELETE or
the 'delete' part of ROW_T_UPDATE. (The 'insert' part is executed if
all BLOBs are present for the 'after' image.) In other words, we must
implement 'replace' semantics for ROW_T_UPDATE.

It is necessary to include the DB_ROLL_PTR in the check, in case a
concurrent DML transaction temporarily violates PRIMARY KEY that would
exist as a result of the ALTER TABLE. The duplicate could go
undetected when a missing BLOB causes a ROW_T_INSERT to be skipped or
ROW_T_UPDATE to be interpreted as ROW_T_DELETE. So, when applying
ROW_T_DELETE or the 'delete' part of ROW_T_UPDATE, we must ensure that
we are deleting the equivalent record as the one that was deleted as
part of the concurrent DML operation.

row_log_table_get_pk(): Copy the DB_TRX_ID, DB_ROLL_PTR from the
B-tree page, instead of making the tuple point to the B-tree page.
(There was a workaround for this bug in row_log_table_delete().)
Add the output parameter "sys" for saving the system columns
for a later row_log_table_delete() invocation.

row_log_table_delete(): Remove the dfield_dup() call for working
around the bug in row_log_table_get_pk(). Remove the parameter "purge"
(assume that the BLOBs can always be eventually purged), and replace
the parameter "trx_id" with "sys". Also log DB_ROLL_PTR.

row_log_table_apply_convert_mrec(): When some BLOBs are missing,
return an incomplete row and set the error to DB_MISSING_HISTORY.
We used to return row=NULL.

row_log_table_apply_insert(): Adapt the logic for the new behaviour of
row_log_table_apply_convert_mrec(). Before and after this change,
we will skip the insert if any BLOBs are missing.

row_log_table_apply_delete(): Replace the parameter new_table with
log, so that we can access "log" in a debug assertion. Add comments.
Expect the log record to contain DB_ROLL_PTR as well.
In case of DB_TRX_ID,DB_ROLL_PTR mismatch, assert that we are
executing ADD PRIMARY KEY.

row_log_table_apply_update(): Adapt the logic for the new behaviour of
row_log_table_apply_convert_mrec(). If there are missing BLOBs,
interpret the ROW_T_UPDATE as ROW_T_DELETE. Before modifying the data,
check that the DB_TRX_ID,DB_ROLL_PTR are matching. Like with ROW_T_DELETE,
a mismatch is only possible in the case of ADD PRIMARY KEY.

row_undo_mod_clust_low(): Add the parameter "sys" for passing the system
columns to row_log_table_delete().

row_undo_mod_clust(): Pass "sys" to row_log_table_delete() as it was
immediately before the update vector was applied.

rb#4280 approved by Jimmy Yang