Bug#18047865 UNDERSIZED MAX-THREAD-INSTANCES CAUSE OVERHEAD IN 5.6 OR CRASH IN 5.7
When tables are re opened from the table cache,
and when the current thread is not instrumented
for the performance schema, the following happens:

- In 5.7, the code crash, while attempting to populate
column OWNER_THREAD_ID in table performance_schema.table_handles

- In 5.6, the code instruments a table handle when not necessary,
causing performance degradations.

The fix is to never instrument table handles,
when re opening a table from the table cache
from a non instrumented thread.