WL#6120- Change master without stopping slave
The test has temporary tables open when stop
slave is executed. After this WL, there will
be a warning in SBR regarding this, but in RBR
there will be no such warning. So the test
starts to produce different result files in SBR
and RBR.

In an earlier patch, the test was made different
for SBR and RBR with an auxiliary test and two
result files.

In this patch we reverted this and instead disabled
warnings at stop slaves where the new warning appears.
This saves us from running the test twice as in the
previous solution.