WL7342 - Refactor the Tablespace class for WL#6205
The Tablespace class was created for the Temporary Tablespace and is
used to refer to the two system tablespaces; innobase_data_file
(IBDATA1 and innobase_temp_data_file (IBTMP1).
This patch refactors that class into three classes;  Datafile,
Tablespace, and SysTablespace.  Tablespace is made into a base class
for SysTablespace.  

These classes are moved from;
  srv/srv0space.cc
  include/srv0space.h
to;
  fsp/fsp0space.cc 
  fsp/fsp0sysspace.cc
  include/fsp0space.h
  include/fsp0sysspace.h
  include/fsp0file.h.

This patch is a convenient way to collect all the non-functional
changes in wl7342 so that they can be reviewed and pushed separately.
There are no function changes, just code reorganization and cleanup.
As part of the review process other non-functional improvements were
made here including handling out-of-memory situations if ut_malloc()
were to return NULL instead of asserting after 60 tries like it does
now.  Several functions that make a filename were consolidated into
fil_make_filepath() and a unit test is created for it.  Several test
cases are cleaned up and new messages are suppressed.

Approved by Sunny in RB#2974