Bug#16767011 - ROW INCORRECTLY NOT UPDATED BY MULTI-TABLE UPDATE
Problem:-
Multiple table updates do not update under certain conditions. 

Analysis:
In our multi update query, with join order a*b*c. We will select 
first join table (which is 'a' here, which here I am calling as
main-table) as the table to be updated at the time of fetching, 
on the basis of some condition(which is checked in 
safe_update_on_fly()) and for all other table we make temporary tables.


In this bug, when one of the field of main-table is 
going to be updated and that field is in where condition, 
with some other field will cause incorrect result.

For example:
CREATE TABLE t1 (c1 INTEGER , c2 INTEGER );
CREATE TABLE t2 (c1 INTEGER , c2 INTEGER);
INSERT INTO t1 VALUES(1,1),(1,4),(1,5);
INSERT INTO t2 VALUES(11,1),(13,1),(15,1);
UPDATE t1 JOIN t2
SET 
t1.c1=30, 
t2.c2=40
WHERE t1.c1=t2.c2 ;

here first, we fetch the value of t1(1,1) and then the value of 
t2(11,1) and compare it. If it matches we change the value of t1 
from (1,1) to (1,30) and for t2 we will change the value in 
temp_table(11,40). Later when we fetch the value of t2 
again(that is 13,1) and compare to t1(that is 1,30).it doesn't 
match, so we can't update it.

So,
SELECT * FROM t2;
c1	c2
11	40
13	1	<--value doesn't change

Solution:-
In safe_update_on_fly(), check if any field of main-tables is going 
to be updated and that field is in where condition with any other 
field in where clause, don't use that table as a main table and make 
a temporary table for it.