Bug#17320977 FAILING ASSERTION: LOCK_REC_HAS_TO_WAIT_IN_QUEUE(LOCK)
Problem:

The function row_vers_impl_x_locked_low() wrongly identifies a transaction to
hold an implicit x lock on a secondary index, which is orphaned.  Orphaned
means that the secondary index is delete marked and purgeable and does not have
its associated clustered index record.  

Solution:

Added an extra check in row_vers_impl_x_locked_low() to detect this situation
and confirm that the transaction does not hold an implicit x lock on the
record.  The flag LOCK_CONV_BY_OTHER is not needed anymore.  This is removed. 

rb#3982 approved by Marko.