Bug #18058884 I_INNODB.INNODB-DOUBLE-WRITE FAILS SPORADICALLY ON 5.6+
Problem:

The test case i_innodb.innodb-double-write is failing sporadically in pb2 for
mysql-5.6 and mysql-trunk.  The error notes that the page 0 of the user
tablespace is not there in the double write buffer.  The root cause of this
sporadic failure is because of the flushing of pages related to undo logs.
These pages can overwrite the page 0 of the user tablespace in the double write
buffer.  

Solution:

Stop the purge thread in the beginning of the test case.  This will ensure that
the undo logs are not freed and hence will not flood the double write buffer.  

rb#4357 in review by Yasufumi.