ndb - disable optimization of Ndbfs::readWriteRequest only for gcc 4.8 and up
Bug #18122881	 REVERT FIX FOR BUG #18055285 FOR LOWER VERSIONS (<4.8) OF GCC

GCC_VERSION was not always defined before include of Ndbfs.cpp as in VoidFs.cpp
so also change to test __GNUC__ and __GNUC_MINOR__ that are predefined for gnu
compiler.

note that for non gnu compiler __GNUC__ and __GNUC_MINOR__ will be treated as 0 in arithmetics
and so optimization not disabled.