It's a follow-up patch for the WL#7193 (Decouple THD and st_transactions).
This patch adds the m_ prefix to the data members flags, changed_tables, 
mem_root of the class Transation_ctx. Such renaming was done to remove
compiler warnings in the methods set_unsafe_rollback_flags() and
add_unsafe_rollback_flags() about the fact that an argument name hides 
the data member and also to make the naming of attributes to be uniform.
Also missed DBUG_VOID_RETURN was added into the method
THD::add_changed_table. Absence of DBUG_VOID_RETURN led to a crash
of the server built in debug mode.