Bug#17814492 - INVALID RESULTS FROM SUBQUERY WITH IN CLAUSE
Problem:
Optimizer decided to push down a condition, when the index
does not have the keypart present in the condition.
      
Analysis:
While checking for possible keys to be considered for a query,
optimizer comes up with a list of possible keys on derived table.
In the query given, optimizer thinks it can generate two keys,
one having a,b and one having just a.
Once the possible keys are updated, optimizer also updates the
corresponding flags for the field on which index could be used.
      
While creating the index on derived table, optimizer sees that
the length of the key for key a,b will be more than the length
of MI_MAX_KEY_LENGTH, so it doesn't create the key. But it
doesn't update the flags in field.
      
Later this lets optimizer to think that there is an index which
has both a and b keyparts and thereby pushing the condition
a1.b == 'xxx' while reading from the materialized table.
This eventually results in evaluation against a field in
record0 that has not been read yet, resulting in lost results.
      
Solution:
When optimizer doesn't create the key on derived table, flags that
were enabled while adding the derived key to the derived key list
should be disabled.
      
This is currently done partially in TABLE::use_index. Do it correctly
now.