Bug#17986385:LOG-WARNINGS=2 UNUSABLE ON SLAVE WITH FILTERS
Problem:
========
If a slave is configured with replication filters, and
log-warnings=2, every statement which is filtered writes an
entry to the error log:

[Warning] Slave SQL: Could not execute Query event. Detailed
error: Slave SQL thread ignored the query because of
replicate-*-table rules.

For busy servers which generate a lot of statements that get
filtered, the result is an error log which can quickly grow
into the GB range.

Analysis:
========
This warning appears when log_warnings > 1 along with error
ER_SLAVE_IGNORE_TABLE. This causes the error log to grow
very quickly.

Fix:
===
"Error_log_throttle" is used to fix the issue.
"Error_log_throttle" is a time based way to control the
error messages. Throttle is used for errors which are
generated very frequently. For such errors first a
log_throttle is defined and a time window_size is specified.
For example when window is set to 1 min, for every 1 min
an error message will be printed which says that this
particular error has occurred n number of times.