BUG#17920923 - BACKPORT PATCH FOR BUG#14511533 INTO 5.6
Problem:
Server crash was observed when code first checked if 
semisync was enabled without lock, if so it takes the
lock and checks again. If semisync gets disabled
in-between the first and second check, an assert
incorrectly referenced a null pointer for active
transaction which leads to the crash. 
    
Solution: 
The assert is relocated onto a position where
active_tranxs buffer is valid.