Bug#18165088 FAILURE TO SETUP EVENT LISTENER DETECTED BY NDB_RESTORE_COMPAT_COMPRESSION
Problem:
- The ndb_mgmd processes the  "listen event" command, writes the reply "listen
event" to m_output and then adds the new listener to the list of active
listeners. But since m_output is buffered(using BufferedSockOutputStream)
there is a probability that check_socket_listener() is called before the
output is flushed. This caused the client to receive the string "<PING>" instead
of the expected "listen event".
- Actually it's also possible for a new log event (from Ndb_mgmd_event_service::log() ) to
  be written to the socket while pending reply was in the output buffer.

Solution:
- Flush the reply from buffer before adding the new event listener to the list
of active listeners.
- Unlock m_mutex while flushing to avoid holding the mutex while doing
network IO. This is smilar to the other place where m_oputput is flushed and
thus the same logic should apply here.