Bug#18122171 DO NOT ASSUME X86 HAS PAUSE INSTRUCTION: MAY BREAK SERVER ON SOLARIS 10
Problem:
The loader in some Solaris versions has a bug due to which it refuses to
start a binary that has been compiled by GCC and uses __asm__("pause")
with the error:
$ ./mysqld
ld.so.1: mysqld: fatal: hardware capability unsupported: 0x2000  [PAUSE]
Killed
$
even though the CPU does have support for the instruction. Binaries that
have been compiled by GCC and use __asm__("pause") on a non-buggy
Solaris get flagged with a "uses pause" flag and thus they are unusable
if copied on buggy Solaris version.

Solution:
To circumvent this we explicitly disable __asm__("pause") when compiling on Solaris.

Approved by: Vasil(rb#4475)