BUG#17458750: ASSERTION FAILED: THD->LOCK == 0 / MYSQL IS FREEING A               THD THOUGH TRX->N_MYSQL_TA
Analysis:
--------
A query of the following format:
SET PASSWORD:=old_password(''), @a:= (select * from t1 where 1);
causes the server to assert on debug build. On a release build,
the error message below is printed during shutdown:
[ERROR] InnoDB: MySQL is freeing a thd though
trx->n_mysql_tables_in_use is 1 and trx->mysql_n_tables_locked
is 1.

In the query mentioned above, the SELECT statement is executed
first followed by SET PASSWORD. The SET PASSWORD statement
updates the system table 'USER'. The SELECT query locks the
user table 't1 during its execution. However during the execution
of SET PASSWORD, before the system table 'USER' is opened and
locked an assert condition checks whether there are already
existing locks on tables specified in the query. In case the
query has already acquired lock, the server crashes since the
condition of acquiring the locks for all the tables specified
in query at the start of the execution is not met.

Fix:
---

SET PASSWORD:= old_password(''), @a=1 is an incorrect usage.

SET PASSWORD is now a separate statement which cannot be
combined with the regular SET syntax and a syntax error is
reported, if combined.

This patch has made changes to the SQL grammar to disallow
such a syntax.