Bug #18144349 INNODB CANNOT USE THE DOUBLEWRITE BUFFER FOR THE FIRST PAGE OF SYSTEM TABLESPACE
Problem:

When the first page of any of the data files of the system tablespace is
corrupt, even if a clean copy of the same is available in the double write
buffer, the server does not start.

Solution:

When the first page of the data files of the system tablespace is corrupt,
try to restore from the doublewrite buffer.

rb#4583 approved by Kevin.