Bug #16204175 COLLATION NAME MISSING FROM LOG MESSAGES               ABOUT LDML DEFINITION PROBLEMS
Description: When trying to use an invalid user defined 
collation error information that can be found in SHOW
WARNINGS also gets logged to the error log, e.g.
2013-01-22 16:03:02 9130 [ERROR] Shift expected at ''
The error message is missing information about the actual
invalid collation

Analysis: When a collation is defined with invalid grammar
rules, " expected at '' " error is logged to error log. But
collation name is missing in this error message. So for a
DBA looking at the error log it is impossible to tell where
the error actually is . So it is better to add collation
name along with the error message which is logged to the
error log.

Fix: As a fix i added the collation name also to the error
message which is logged to the error log.