Bug#18236000 TRUNCATE TABLE IS WRITING TO SYSTEM TABLESPACE WITHOUT REDO LOGGING
The methods truncate_t::drop_indexes() and
truncate_t::create_indexes() were disabling redo logging when
modifying the system tablespace (ibdata1).

There are 2 possible problems caused by not updating the inode page
(pointed to by BTR_SEG_LEAF or BTR_SEG_TOP) in sync with the
allocation bitmap pages (usually page 0:0):

(1) If the inode page is written out but the bitmap page is not, in
the event of a crash we will leak (never free) the pages.

(2) If the bitmap page was written out but the inode page was not, we
could double-free a page, hitting an assertion failure on crash
recovery.

Redo logging makes these updates atomic (either both pages are updated
or none of them are).

rb#4654 approved by Vasil Dimov and Sunny Bains