Bug#18144349 INNODB CANNOT USE THE DOUBLEWRITE BUFFER FOR THE FIRST PAGE OF SYSTEM TABLESPACE
Follow up fix after pb2 failure.  The debug code innodb_make_page_dirty()
was always dirtying the first page of the tablespace.  Fixed it to dirty
the specified page.  In the test case innodb-double-write.test, the 2nd
page of the t1 tablespace is not dirtied.  Dirty it correctly.  Also,
for each recovery, do not dirty many pages.  It then gets overwritten
in the double write buffer.  Also trying to minimize the interference from
the purge thread.