Bug #18001777 MYSQL SLAVE SQL THREAD MEMORY hogging
Technically it's not a leak which the reporter admits himself, see
the bug report.

Source code examination showed that the problem was in usage of an
incorrect mem-root. The main THD root of Coordinator thread was never
reset.
After discussion with reviewer a solution to dismantle the specific
root is decided. Instead array of pointers to database names of a
maximum size is allocated in the stack which is correct 'cos database
names do not get out scope of gets destructed while their pointers are
processed.
The fixes are verified on memory consumption to prove correctness of
the found suspect and the fixing method.