Bug #18196562: SEGV in NdbApi in NdbReceiver::receive_packed_recattr
An NDBAPI application sends a scan query to a data node, which is 
processed by the TC. The TC forwards the LQHKEYREQ request to the 
appropriate LDM, and aborts the transaction if it does not receive 
a LQHKEYCONF response within the specified time limit. After the 
transaction is successfully aborted, the TC sends a TCROLLBACKREP 
to the NDBAPI client. The NDBAPI client processes the TCROLLBACKREP 
by cleaning up the Ndb objects associated with the transaction.

The NDBAPI client receives the data which it has requested in the 
form of TRANSID_AI signals sent by the LDM. These TRANSID_AI signals 
are buffered for sending at the data node, and may be delivered to 
the NDBAPI client after a delay. On receiving a TRANSID_AI signal, 
the NDBAPI checks the transaction state and ID: if these are as 
expected, it processes the signal using the Ndb objects associated 
with that transaction.

The error scenario is as follows:
- The NDBAPI client sends a request to the TC for data.
- The TC sends an LQHKEYREQ to an LDM.
- The TC times out waiting for the LQHKEYCONF from the LDM.
- A TRANSID_AI is buffered for sending at the data node, but there 
is a delay in sending it.
- The TC aborts the transaction and receives a confirmation from 
the LDM.
- The TC sends a TCROLLBACKREP to the NDBAPI client, which cleans up 
the Ndb objects for that transaction.
- The buffered TRANSID_AI is finally delivered to the NDBAPI client, 
which attempts to process it using the Ndb objects which have already 
been cleaned up, resulting in a segfault.

Added a fix in the NDBAPI, so that the Ndb object cleanup done for 
TCROLLBACKREP now includes invalidation of the transaction ID. So, 
for a given transaction, any signal which is received after the 
TCROLLBACKREP arrives, does not pass the transaction ID check and 
is silently dropped.