Bug #16693068 	7.1.18 TO 7.1.22-27 UPGRADE NOT POSSIBLE IF LARGE NUMBER OF TABLES PRESENT
Older code stored unlimited log part ids in the fragment definition,
and used modulo division at 'runtime' to get the correct log part.
Current code stores modulo log part ids and uses them directly.

During upgrade we must do this conversion if necessary.

One downside of storing log parts in their 'modulo' form is that an
increase in the number of log parts cannot remap existing fragments
to the new log parts.  That is considered out of scope here.

testUpgrade is extended to cover the scenario required here :
 - Many tables
 - Many fragments/table
 - System restart upgrade rather than node restart upgrade

An execution of this upgrade test is added to the upgrade-tests
suite for running in Autotest.