Bug#18069105 - ADD FUSIONIO ATOMIC WRITE SUPPORT FOR LINUX
If any of the system tablespace (ibdata1) files, not the log or UNDO
tablespace files, is on a file system that supports FusionIO atomic writes
then disable the InnoDB double write buffer. It would be more desirable to
make this setting per physical file. This patch implements an all or nothing
approach to keep things simple for now.
			            
This change shouldn't make any difference on non-FusionIO HW. It will only
be enabled on FusionIO HW that supports atomic-writes on Linux.
					        
Approved by Yasufumi rb#4323