Bug #18114294   CRASH IN ITEM_FUNC_LIKE::BM_MATCHES
Problem
-------

While building the bad character shift array
we first typecast the pattern value to unsigned
char and then to unsigned int to determine the
index of the array ,but while doing the search
when bad character shift array is indexed
according to the text (from which pattern is
searched) value,we were directly typecasting
the text value to unsigned int which was
causing the crash.

Fix
---
Typecasting the text value to unsigned char 
is necessary ,typecasting to unsigned int 
is not required.