Bug#18136628: RANGE OPTIMIZER: IMERGE_LIST_OR_TREE() ERROR DUE               TO SHALLOW COPY
imerge_list_or_tree() performs the following OR operation in 
the range optimizer: "<list_of_index_merges> OR <sel_tree>",
e.g

Index merge list:
   (pred1_idx1 OR pred1_idx2)    <- an index merge
   AND
   (pred2_idx1 OR pred2_idx2)    <- another index merge
OR
sel_tree:
   pred3_idx1

as follows: 'sel_tree' is first ORed with 
"(pred1_idx1 OR pred1_idx2)" and then with 
"(pred2_idx1 OR pred2_idx2)" by calls to tree_or(). However,
tree_or() is allowed to modify its parameters, and because both
index merges should be ORed with the same 'sel_tree' predicate,
imerge_list_or_tree() makes a copy of it for each call to 
tree_or(). This is done to avoid the following situation:

Index merge list:
   (col1 < 10 OR <something>)
   AND
   (col1 = 5 OR <something>)
OR
sel_tree:
   col1 > 9

imerge_list_or_tree() does:
1) Perform "col1 < 10 OR col1 > 9" => always true, 'sel_tree' 
   is marked accordingly
2) Perform "col1 = 5 OR true" => always true since 'sel_tree' 
   was marked as true. This is obviously incorrect and was 
   caused by step 1) modifying 'sel_tree'. The remedy is to
   copy 'sel_tree'.

Without copying 'sel_tree', the range would therefore be:

   (TRUE OR <something>) AND (TRUE OR <something) 
=> TRUE
 
While the correct range is:

   (TRUE OR <something>) 
   AND ((col1 = 5 OR col1 > 9) OR <something>)
=> ((col1 = 5 OR col1 > 9) OR <something>)

The problem in this bug was that while imerge_list_or_tree() 
correctly made a copy of 'sel_tree' before each OR, the copy 
it made was shallow. Results from step 1) therefore still 
affected the copy of 'sel_tree' used in step 2.

The fix is to change the SEL_TREE ctor to make a deep 
copy. The ctor is not used anywhere else, so this only 
affects imerge_list_or_tree()