Fix for Bug#18293112
  NDB.HPP DEPENDS ON PORTLIB/NDBTICK.HPP THAT'S NOT INCLUDED IN BINARY DIST

This patch fix a regression introduced by the refactoring part of
bug #17647637: 'Refactor of NdbTick as prep for real fix.'

5252 Ole John Aske	2013-11-18
Part1 of fix for bug#17647637: Refactor of NdbTick as prep for real fix.

Thix patch reverts the introduction of 'NDB_TICKS' and 'portlib/NdbTick.h'
in the API header files.