BUG#17632285 SLAVE CAN'T CONTINUE REPLICATION AFTER MASTER'S              CRASH RECOVERY
Binary events might be sent to slaves before they are flushed
to disk on master, even sync_binlog is set to 1. It can cause
two problems if the master restarts after an OS crash.
* Replication cannot continue because the slaves are
  requesting to replication the events don't exist on master.
* Data exists on slaves, but not exists on the master.

The problems are expected on less durable settings(
sync_binlog != 1), but it should not happen on durable
setting(sync_binlog = 1).

Since 5.6 binlog group commit implementation, binlog write
and sync have been protected by separate mutexes. So dump
threads can read the binary events simultaneously or even
before it is synced to disk.

To fixing the problem on durable setting, LOCK_log is hold
in sync stage and it is released after the binary events are
synced to disk.